+++ .editorconfig
@@ -9,5 +9,8 @@ trim_trailing_whitespace = true
 insert_final_newline = true
 max_line_length = 100
 
+[tsconfig*.json]
+insert_final_newline = true
+
 [*.md]
 trim_trailing_whitespace = false
