+++ packages/scribe/prettier.config.js
@@ -1,5 +1,7 @@
 const config = {
   plugins: ["prettier-plugin-tailwindcss"],
+  endOfLine: "lf",
+  insertFinalNewline: true,
 };
 
 export default config;
