+++ packages/scribe/src/components/NodesMenu/Menu/MenuContext.tsx
@@ -1,17 +1,6 @@
-/* eslint-disable react-refresh/only-export-components */
-import { createContext, useContext } from "react";
+import { createContext } from "react";
 import { useMenuCore } from "./useMenuCore";
 
 type MenuContextType = ReturnType<typeof useMenuCore>;
 
-const MenuContext = createContext<MenuContextType | undefined>(undefined);
-
-export function useMenuContext() {
-  const context = useContext(MenuContext);
-  if (!context) {
-    throw new Error("useMenuContext must be used within a MenuProvider");
-  }
-  return context;
-}
-
-export { MenuContext };
+export const MenuContext = createContext<MenuContextType | undefined>(undefined);
