+++ packages/scribe/src/components/NodesMenu/Menu/useMenuContext.tsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { MenuContext } from "./MenuContext";
+
+export function useMenuContext() {
+  const context = useContext(MenuContext);
+  if (!context) {
+    throw new Error("useMenuContext must be used within a MenuProvider");
+  }
+  return context;
+}
